Controller manager (#68)
* try and improve controller management

* simplify PnPDetails

* still weird things happening

* more work, still somehow getting mixed results where two devices are on the same slot...

* more work

* fix broken check on XUsbDeviceRemoved

* few tweaks

* hacking

* not sure I can improve past this point

* is it really doing anything ?

* might not cause any harm

* Try now

* could it be it !?

* (re)implement ControllerManagement toggle

* fix LegionGo LED control

* force gamepadmode to XBOX on Legion Go when XInput controller lands

+++ HandheldCompanion/Controllers/DInputController.cs
@@ -17,7 +17,7 @@ public class DInputController : IController
             return;
 
         this.joystick = joystick;
-        UserIndex = joystick.Properties.JoystickId;
+        UserIndex = (byte)joystick.Properties.JoystickId;
 
         if (details is null)
             return;
